t test per cpg by different cancer types

+++ ttest_on_CpG_by_cancer.py
@@ -22,29 +22,30 @@ pvalue_dir = "/data2/external_data/Sun_Zhifu_zxs01/summerprojects/ltian/MethylDB
 
 # FUNCTIONS
 def perform_t_test(cpg_id):
-    table = "<table class='table table-hover'><thead><tr><th>Cancer type</th><th>p value</th></tr></thead><tbody>"
-    pvals = ""
-    datafile = cpg_dir + cpg_id + ".txt"
-    df = pd.read_table(datafile,sep=",",header=0,index_col=None)
-    cancer_counts = {}
-    for i in df["Acronym"]:
-        if i not in cancer_counts:
-            cancer_counts[i] = 1
-        else:
-            cancer_counts[i] += 1
-    cancer_to_test = [i for i in cancer_counts if cancer_counts[i]>100 and df[(df["Acronym"]==i) & (df["TumorNormal"]=="Normal")].shape[0]>3]
-    for i in cancer_to_test:
-        table += "<tr><td scope='col'>{0}</td>".format(i)
-        subdf = df[df["Acronym"]==i]
-        t,p = ttest_ind(subdf[subdf["TumorNormal"]=="Tumor"]["Value"], subdf[subdf["TumorNormal"]=="Normal"]["Value"],equal_var=False)
-        table += "<td>{0}</td>".format(p)
-        table += "</tr>"
-        pvals += "{0},{1}\n".format(i,p)
-    table += "</tbody></table>"
-    with open(pvalue_table_dir + cpg_id + ".html","w") as f:
-        f.write(table)
-    with open(pvalue_dir + cpg_id + ".txt","w") as f:
-        f.write(pvals)
+    if os.path.isfile(cpg_dir+cpg_id+".txt"):
+        table = "<table class='table table-hover'><thead><tr><th>Cancer type</th><th>p value</th></tr></thead><tbody>"
+        pvals = ""
+        datafile = cpg_dir + cpg_id + ".txt"
+        df = pd.read_table(datafile,sep=",",header=0,index_col=None)
+        cancer_counts = {}
+        for i in df["Acronym"]:
+            if i not in cancer_counts:
+                cancer_counts[i] = 1
+            else:
+                cancer_counts[i] += 1
+        cancer_to_test = [i for i in cancer_counts if cancer_counts[i]>100 and df[(df["Acronym"]==i) & (df["TumorNormal"]=="Normal")].shape[0]>3]
+        for i in cancer_to_test:
+            table += "<tr><td scope='col'>{0}</td>".format(i)
+            subdf = df[df["Acronym"]==i]
+            t,p = ttest_ind(subdf[subdf["TumorNormal"]=="Tumor"]["Value"], subdf[subdf["TumorNormal"]=="Normal"]["Value"],equal_var=False)
+            table += "<td>{0}</td>".format(p)
+            table += "</tr>"
+            pvals += "{0},{1}\n".format(i,p)
+        table += "</tbody></table>"
+        with open(pvalue_table_dir + cpg_id + ".html","w") as f:
+            f.write(table)
+        with open(pvalue_dir + cpg_id + ".txt","w") as f:
+            f.write(pvals)
 
 # MAIN
 if __name__ == '__main__':
